Function to get available tickers from downloaded zip file

This function will read the zip file downloaded from Bovespa and output a numeric vector where the names of the elements represents the different tickers and the numeric values as the number of trades for each ticker

ghfd_get_available_tickers_from_file(out.file)

Arguments

out.file
Name of downloaded file with HFT data from Bovespa

Value

A dataframe with the number of trades for each ticker found in file

Examples

## get file from package (usually this would be been downloaded from the ftp)
out.file <- system.file("extdata", 'NEG_OPCOES_20151126.zip', package = "GetHFData")

df.tickers <- ghfd_get_available_tickers_from_file(out.file)

print(head(df.tickers))

#>   tickers n.obs
#> 1 PETRL80  2882
#> 2  PETRL9   985
#> 3 VALEL14   754
#> 4 VALEL43   679
#> 5 PETRL70   514
#> 6 PETRX80   507
#>                                                                  f.name
#> 1 /home/msperlin/GitRepo/GetHFData/inst/extdata/NEG_OPCOES_20151126.zip
#> 2 /home/msperlin/GitRepo/GetHFData/inst/extdata/NEG_OPCOES_20151126.zip
#> 3 /home/msperlin/GitRepo/GetHFData/inst/extdata/NEG_OPCOES_20151126.zip
#> 4 /home/msperlin/GitRepo/GetHFData/inst/extdata/NEG_OPCOES_20151126.zip
#> 5 /home/msperlin/GitRepo/GetHFData/inst/extdata/NEG_OPCOES_20151126.zip
#> 6 /home/msperlin/GitRepo/GetHFData/inst/extdata/NEG_OPCOES_20151126.zip